Use `rustc-hex v2.0` with explicit types
The issue was that the trait `FromHex` has changed and is now using
`FromIterator<u8>` instead of `FromIterator<Vec<u8>>` and was `inferred`
to `[u8]` instead of `Vec<u8>`. Thus, has not compile-time size and
won't work!

Well, this could be improved to be stack allocated because
we know the `size` of the hexstring but in this case I think
is sufficient and only used in tests!