Allow pubkey recovery for all-zero messages (#369)
* Allow pubkey recovery for all-zero messages

After https://github.com/openethereum/openethereum/pull/11406 it is no longer possible to to public key recovery from messages that are all-zero. This create issues when using the `ecrecover` builtin because externally produced signatures may well provide a message (i.e. a preimage) that is all-zeroes.
This works around the problem at the cost of cloning the incoming message and create a `ZeroesAllowedMessage` wrapper around it. The `ZeroesAllowedMessage` implements the `ThirtyTwoByteHash` trait from `rust-secp256k1` which circumvents the zero-check.

In a follow-up PR we'll likely change the interface of `recover()` to take a `ZeroesAllowedMessage` directly, thus removing the unneeded clone.

* Inner doesn't need to be pub

* Update parity-crypto/src/publickey/ecdsa_signature.rs

Co-Authored-By: Andronik Ordian <write@reusable.software>

* Docs and review grumbles

* Add `recover_allowing_all_zero_message()`
Revert `recover()` to previous behaviour: no zero-messages allowed
Docs and cleanup

* Obey the fmt

Co-authored-by: Andronik Ordian <write@reusable.software>