Add automatic methods to KeyValueDB which test for the existence of certain objects (#409)
Motivation: We have a use case in which we care about a large object's
existence, but don't care to load megabytes of data from disc. We have
only a `dyn KeyValueDB` handle available; whatever we use has to be
part of the trait.

By adding automatic implementations for the new methods, we avoid
breaking existing code which is no worse than the existing strategy.

Part 2 of this fix will be tracking down the concrete type in use
and implementing specializations of those methods, so that we actually
do save work, but that's dependent on this PR being merged.